JAMES-2525 Extract fake properties provider to common package
the fake properties provider was deemed useful enough in code review to
be raised from its internal class status to a  public common component.
